+++ pydidas/gui/frames/quick_integration_frame.py
@@ -113,6 +113,9 @@ class QuickIntegrationFrame(BaseFrame):
                 _generic.params, diffraction_exp=self._EXP
             ),
         }
+        self._image = None
+        self._bc_controller = None
+        self._roi_controller = None
 
     def build_frame(self):
         """
@@ -181,6 +184,9 @@ class QuickIntegrationFrame(BaseFrame):
             _w.io_edited.connect(partial(self._update_xray_param, _param_key, _w))
         for _param_key in ["beamcenter_x", "beamcenter_y"]:
             self.param_widgets[_param_key].io_edited.connect(self._update_beamcenter)
+        self.param_widgets["detector_mask_file"].io_edited.connect(
+            self._new_mask_file_selection
+        )
 
     @QtCore.Slot(str, object)
     def open_image(self, filename: Union[str, Path], kwargs: dict):
@@ -328,12 +334,38 @@ class QuickIntegrationFrame(BaseFrame):
         _det_model = self.get_param_value("detector_model")
         if _det_model == "Custom detector":
             _pxsize = self._config["custom_det_pxsize"]
+            self._config["detector_name"] = None
             self.set_param_value("detector_name", "Custom detector")
+            _func = self._bc_controller.set_mask_file
         else:
             _det_name = _det_model.split("]")[1].strip()
             self._EXP.set_detector_params_from_name(_det_name)
             _pxsize = self.get_param_value("detector_pxsizex")
+            self._config["detector_name"] = _det_name
+            _func = self._bc_controller.set_new_detector_with_mask
+        if not self.get_param_value("detector_mask_file").is_file():
+            _func(self._config["detector_name"])
         self._update_detector_pxsize(_pxsize)
+
+    @QtCore.Slot(str)
+    def _new_mask_file_selection(self, mask_filename: str):
+        """
+        Propagate the new mask to the beamcente controller.
+
+        Parameters
+        ----------
+        mask_filename : str
+            The name of the new mask file or None to disable mask usage.
+        """
+        _path = Path(mask_filename)
+        if _path.is_file():
+            self._bc_controller.set_mask_file(_path)
+        elif self._config["detector_name"] is not None:
+            self._bc_controller.set_new_detector_with_mask(
+                self._config["detector_name"]
+            )
+        else:
+            self._bc_controller.set_mask_file(None)
 
     @QtCore.Slot(str)
     def _update_beamcenter(self, _):
@@ -361,7 +393,6 @@ class QuickIntegrationFrame(BaseFrame):
         self._roi_controller.toggle_marker_color_param_visibility(not _active)
         self._widgets["file_selector"].setEnabled(not _active)
         self._roi_controller.toggle_enable(not _active)
-        self._bc_controller.selected_points
         if _active:
             self._bc_controller.show_plot_items("all")
             self._roi_controller.remove_plot_items("roi")